инкремент 10: подключение к БД
