+++ go.mod
@@ -4,6 +4,7 @@ go 1.21.10
 
 require (
 	github.com/go-chi/chi/v5 v5.0.12
+	github.com/jackc/pgx/v5 v5.6.0
 	github.com/mailru/easyjson v0.7.7
 	github.com/stretchr/testify v1.9.0
 	go.uber.org/zap v1.27.0
@@ -11,8 +12,13 @@ require (
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
 	github.com/josharian/intern v1.0.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rogpeppe/go-internal v1.12.0 // indirect
 	go.uber.org/multierr v1.10.0 // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
